dev-mapping-1

+++ src/lvio_fusion/src/frontend.cpp
@@ -194,7 +194,7 @@ bool Frontend::InitFramePoseByPnP()
 int Frontend::TrackLastFrame()
 {
     // use LK flow to estimate points in the last image
-    std::vector<cv::Point2d> kps_last, kps_current;
+    std::vector<cv::Point2f> kps_last, kps_current;
     std::vector<visual::Landmark::Ptr> landmarks;
     for (auto pair_feature : last_frame->features_left)
     {
@@ -272,7 +272,7 @@ int Frontend::DetectNewFeatures()
         cv::rectangle(mask, feature->keypoint - cv::Point2d(10, 10), feature->keypoint + cv::Point2d(10, 10), 0, cv::FILLED);
     }
 
-    std::vector<cv::Point2d> kps_left, kps_right;
+    std::vector<cv::Point2f> kps_left, kps_right;   // must be point2f
     cv::goodFeaturesToTrack(current_frame->image_left, kps_left, num_features_ - current_frame->features_left.size(), 0.01, 30, mask);
 
     // use LK flow to estimate points in the right image
